+++ src/Plutarch/Extra/ScriptContext.hs
@@ -57,7 +57,7 @@ import Plutarch.Api.V2 (
 import Plutarch.Extra.AssetClass (PAssetClass, passetClassValueOf)
 import Plutarch.Extra.Function ((#.*))
 import Plutarch.Extra.Functor (PFunctor (pfmap))
-import Plutarch.Extra.List (pfirstJust)
+import Plutarch.Extra.List (pfindJust)
 import Plutarch.Extra.Maybe (pfromJust, pisJust, pjust, pnothing)
 import Plutarch.Extra.TermCont (pletC, pmatchC)
 import Plutarch.Unsafe (punsafeCoerce)
@@ -170,7 +170,7 @@ pisTokenSpent =
 pfindTxInByTxOutRef :: forall (s :: S). Term s (PTxOutRef :--> PBuiltinList PTxInInfo :--> PMaybe PTxInInfo)
 pfindTxInByTxOutRef = phoistAcyclic $
     plam $ \txOutRef inputs ->
-        pfirstJust
+        pfindJust
             # plam
                 ( \r ->
                     pmatch r $ \(PTxInInfo txInInfo) ->